requirements.txt
langgraph>=0.0.42
langchain>=0.1.16
langchain-core>=0.1.52
langchain-openai>=0.1.6
langchain-mcp-adapters>=0.1.0
mcp>=0.2.0
openai>=1.30.1
python-dotenv>=1.0.1
nest_asyncio>=1.6.0
